Move SimplePie class to namespace
